Change size_t to int to avoid SAIS conversions

+++ FileManipulation.cpp
@@ -28,9 +28,9 @@ std::shared_ptr<std::string> Load_Sequence(const char *filename){
     return std::make_shared<std::string>(sequenceString);
 }
 
-std::shared_ptr<std::string> Combine_Sequences(const std::vector<std::shared_ptr<std::string>> &seqStringArray, size_t numSequences) {
+std::shared_ptr<std::string> Combine_Sequences(const std::vector<std::shared_ptr<std::string>> &seqStringArray, int numSequences) {
     char sentinel = '!'; //Sentinel not included in language dictionary.
-    size_t index;
+    int index;
     std::string combinedString;
 
     for(index = 0; index < numSequences; ++index){
@@ -43,10 +43,10 @@ std::shared_ptr<std::string> Combine_Sequences(const std::vector<std::shared_ptr
 }
 
 bool Write_Matches(std::unordered_map<std::string, MatchLocations> &matchesMap,std::vector<double> &similarityMetrics,
-                   size_t &numSequences, const std::string &outFilename){
+                   int &numSequences, const std::string &outFilename){
     std::ofstream File;
     File.exceptions(~::std::ios_base::goodbit);
-    std::vector<std::unordered_set<size_t>> matchIndicesVector;
+    std::vector<std::unordered_set<int>> matchIndicesVector;
 
     try{
         //Clear file before writting to file
@@ -56,7 +56,7 @@ bool Write_Matches(std::unordered_map<std::string, MatchLocations> &matchesMap,s
         File.open(outFilename); //Open for writing
 
         //Write Similarity Metrics
-        size_t i;
+        int i;
         for (i = 0; i < numSequences; ++i) {
             File << similarityMetrics[i] << std::endl;
         }
