+++ PossibleMatches.h
@@ -12,15 +12,15 @@
 
 class PossibleMatches {
 private:
-    std::unordered_map<int, size_t> SAIndexMap; //Key : SAIndex, Val: seqIndex
-    std::unordered_set<size_t> uniqueSeqIndices;
+    std::unordered_map<int, int> SAIndexMap; //Key : SAIndex, Val: seqIndex
+    std::unordered_set<int> uniqueSeqIndices;
 public:
     PossibleMatches();
-    void InsertMatch(int SAIndex, size_t &seqIndex);
-    void AddSubMatch(std::unordered_map<int, size_t> & map);
-    std::unordered_map<int, size_t> & getSAIndexMap();
-    size_t UniqueSetIndices();
-    std::shared_ptr<std::vector<std::unordered_set<size_t>>> TransferMatches();
+    void InsertMatch(int SAIndex, int &seqIndex);
+    void AddSubMatch(std::unordered_map<int, int> & map);
+    std::unordered_map<int, int> & getSAIndexMap();
+    int UniqueSetIndices();
+    std::shared_ptr<std::vector<std::unordered_set<int>>> TransferMatches();
 };
